requirements.txt
alembic==1.4.2
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
bs4==0.0.1
cffi==1.14.2
click==7.1.2
cryptography==3.1
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-slugify==4.0.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
text-unidecode==1.3
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
